+++ crates/egui-wgpu/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the `egui-wgpu` integration will be noted in this file.
 
 ## Unreleased
 * Return `Err` instead of panic if we can't find a device ([#2428](https://github.com/emilk/egui/pull/2428)).
+* `winit::Painter::set_window` is now `async` ([#2434](https://github.com/emilk/egui/pull/2434)).
 
 
 ## 0.20.0 - 2022-12-08 - web support